+++ CodeChat_Server/CodeChat_Server/CodeChat_Client/static/splitter.js
@@ -161,6 +161,7 @@ var splitMe = {
 
 
 
-addEvent(window,'load',function () {
-    splitMe.init();
-});
+// BAJ: invoke this manually, since the window load event happens after the CodeChat client initialized and uses the splitter.
+//addEvent(window,'load',function () {
+//    splitMe.init();
+//});