+++ docs/CHANGELOG.rst
@@ -13,7 +13,7 @@ Change Log
 **********
 -   `Github master <https://github.com/bjones1/CodeChat_system.git>`_:
 
-    -   No changes yet.
+    -   Correctly size splitter on startup.
 
 -   0.0.16, 18-Jan-2021: